requirements.txt
streamlit
llama-cpp-python
requests
beautifulsoup4
nltk
numpy
pandas
scikit-learn
